requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20

# ─── Pydantic ─────────────────────────────────────────────────────────────────
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0

# ─── Google Cloud ─────────────────────────────────────────────────────────────
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-aiplatform==1.140.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-documentai==3.10.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.65.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0

# ─── Data / parsing ───────────────────────────────────────────────────────────
numpy==2.2.6
pandas==2.3.3
pyarrow==23.0.1
openpyxl==3.1.5
python-docx==1.1.2
pymupdf>=1.24.0
db-dtypes==1.5.0

# ─── HTTP / async ─────────────────────────────────────────────────────────────
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
h11==0.16.0
websockets==16.0
requests==2.32.5

# ─── Rate limiting ────────────────────────────────────────────────────────────
slowapi==0.1.9
limits==5.8.0

# ─── Crypto / auth ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
email-validator==2.2.0
resend==2.26.0
bcrypt==5.0.0
cryptography==46.0.5
cffi==2.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
rsa==4.9.1

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0
pytz==2026.1.post1
tzdata==2025.3
tenacity==9.1.4
packaging==26.0
six==1.17.0
distro==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
sniffio==1.3.1
proto-plus==1.27.1
protobuf==6.33.5
pycparser==3.0
docstring_parser==0.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
exceptiongroup==1.3.1
python-dotenv>=1.0.0
